llm-doc-loader/requirements.txt
paddlepaddle
paddleocr
torch
torchvision
transformers
accelerate
pillow
opencv-python
huggingface_hub
fastapi
uvicorn
python-multipart
protobuf==3.20.*
openai
python-stdnum